More robust `_set_prometheus_client_use_created_metrics`
