+++ synapse/metrics/__init__.py
@@ -25,6 +25,7 @@ import logging
 import os
 import platform
 import threading
+from importlib import metadata
 from typing import (
     Callable,
     Dict,
@@ -41,6 +42,7 @@ from typing import (
 )
 
 import attr
+from pkg_resources import parse_version
 from prometheus_client import CollectorRegistry, Counter, Gauge, Histogram, Metric
 from prometheus_client.core import (
     REGISTRY,
@@ -80,6 +82,48 @@ We're purposely not using the `instance` label for this purpose as that should b
 terms, an endpoint you can scrape is called an *instance*, usually corresponding to a
 single process." (source: https://prometheus.io/docs/concepts/jobs_instances/)
 """
+
+CONTENT_TYPE_LATEST = "text/plain; version=0.0.4; charset=utf-8"
+
+
+def _set_prometheus_client_use_created_metrics(new_value: bool) -> None:
+    """
+    Sets whether prometheus_client should expose `_created`-suffixed metrics for
+    all gauges, histograms and summaries.
+
+    There is no programmatic way in the old versions of `prometheus_client` to disable
+    this without poking at internals; the proper way in the old `prometheus_client`
+    versions (> `0.14.0` < `0.18.0`) is to use an environment variable which
+    prometheus_client loads at import time. For versions > `0.18.0`, we can use the
+    dedicated `disable_created_metrics()`/`enable_created_metrics()`.
+
+    The motivation for disabling these `_created` metrics is that they're a waste of
+    space as they're not useful but they take up space in Prometheus. It's not the end
+    of the world if this doesn't work.
+    """
+    import prometheus_client.metrics
+
+    if hasattr(prometheus_client.metrics, "_use_created"):
+        prometheus_client.metrics._use_created = new_value
+    # Just log an error for old versions that don't support disabling the unecessary
+    # metrics. It's not the end of the world if this doesn't work as it just means extra
+    # wasted space taken up in Prometheus but things keep working.
+    elif parse_version(metadata.version("prometheus_client")) < parse_version("0.14.0"):
+        logger.error(
+            "Can't disable `_created` metrics in prometheus_client (unsupported `prometheus_client` version, too old)"
+        )
+    # If the attribute doesn't exist on a newer version, this is a sign that the brittle
+    # hack is broken. We should consider updating the minimum version of
+    # `prometheus_client` to a version (> `0.18.0`) where we can use dedicated
+    # `disable_created_metrics()`/`enable_created_metrics()` functions.
+    else:
+        raise Exception(
+            "Can't disable `_created` metrics in prometheus_client (brittle hack broken?)"
+        )
+
+
+# Set this globally so it applies wherever we generate/collect metrics
+_set_prometheus_client_use_created_metrics(False)
 
 
 class _RegistryProxy:
